Move time-related utils from util.h to time-util.h
